Remap the ports returned by the Container in the ProtocolMetaData
The Container only knows about the ports it self is bound to,
not how Docker is remapping them from the outside.

When e.g. the Container is bound to port 8080 inside Docker
and Docker PortForward from external port 8081 to internal
8080 the ProtocolMetaData is of. This effects the ServletProtocol
which is attempting to use this information to communicate with
the in-container test case as well as URL injection on the client
side via @ArquillianResource.

When the ProtocolMetaData is created, the object is intercepted
bu Cube to update the data based on the port we know are being
forwarded.

fixes #12