Add support for auto starting Cubes in parallel
Cubes configured for auto start will now be sorted and
executed in correct start up order based on the
defined links in the configuration.

When linked levels are defined, A->B->C, the parallel execution
will occure in steps per link.

If A->B:
 * then no real parallel can be done since B needs to be
   started before A

If A->B and C->B
 * then B will be started, so A and C in parallel.

If A and B (no link)
 * then A and B will be started in parallel

If A->B and autoStartContainers= A
 * then B will also be started since it's required by A

fixes #73