Add support for OpenShift PortForward
This allow to communicate with ports not bound on a external network.
The communication happen via a PortForward on the OpenShift host. After the
forward is setup, a local port is used to proxy to the internal container via the host.

localhost:8080 -> host:forward-port -> container:local-port

Static configuration is updated to connect to localhost:8080.

fixes #272 #261 #263
resolves #259