resolves issue #310 by rebuilding dockerClient after a Docker build command
Actually when trying to build an image with Arquillian Cube, docker host throwed
the exception message: http: Hijack is incompatible with use of CloseNotifier.
It seems to be a problem with go and should be fixed in go 1.6 (and maybe in
Docker 1.11.0).
For now, closing and recreating the Docker client after a call to the build Docker
command seems to solves the problem.