PortForwarder: Make sure to release the error stream
There is a race condition at the time we try to read
from the error stream: As we do it in a worker thread,
when the thread is executed, this stream may be already
closed by the remote endpoint. So, sometimes the callbacks
for StringReadChannelListener were never called, resulting
in the errorComplete latch never being decreased, causing
a hanging in the client side.

Closes #449.