PortForwarder: Use the IO thread in the response listener
Instead of the worker one.

This fixes a deadlock that happens when a number higher than
WORKER_TASK_MAX_THREADS (currently 32) of clients connect
"at the same time". Then we reach the maximum of worker threads
for incoming connections. At this point no worker thread is
available to handle this IO, pushing it to the worker queue.

Because this IO doesn't happen, communication (client server chat
and eventually a disconnect) also doesn't happen. If a disconnect
doesn't happen, the thread will never finish their job, thus never
giving room in the worker pool. DEADLOCK.