Improvements on Port Forwader (#450)
* PortForwarder: Remove workaround for UNDERTOW-607

This is fixed upstream since 1.3.13.

* PortForwarder: Removed unused variables, imports and comments

* PortForwarder: Remove flushRequest() method

It's not necessary anymore.

* PortForwarder: Make sure to release the error stream

There is a race condition at the time we try to read
from the error stream: As we do it in a worker thread,
when the thread is executed, this stream may be already
closed by the remote endpoint. So, sometimes the callbacks
for StringReadChannelListener were never called, resulting
in the errorComplete latch never being decreased, causing
a hanging in the client side.

Closes #449.

* PortForwarder: Use the IO thread in the response listener

Instead of the worker one.

This fixes a deadlock that happens when a number higher than
WORKER_TASK_MAX_THREADS (currently 32) of clients connect
"at the same time". Then we reach the maximum of worker threads
for incoming connections. At this point no worker thread is
available to handle this IO, pushing it to the worker queue.

Because this IO doesn't happen, communication (client server chat
and eventually a disconnect) also doesn't happen. If a disconnect
doesn't happen, the thread will never finish their job, thus never
giving room in the worker pool. DEADLOCK.

Closes #448.