Kubernetes/Openshift: Defaults "namespace.cleanup.enabled" to true
We should keep the default value as "true" regardless of whether we
are going to destroy the namespace or not. This is to keep consistency
when users call "configuration.isNamespaceCleanupEnabled()".

Users of this library can rely on the above method to decide when
to cleanup stuff they have created. So, the library must be
consistent on their flags and documentation.

Closes #747.