Kubernetes: Add the ability to capture pods logs
It might be useful for debugging purposes to have the ability
to automatically capture console log for all pods created
during the test execution.

It is disabled by default. To enable it, a new boolean flag is
added: "logs.copy".

The logs will be stored in the local filesystem, one file per pod.
Filenames will be "ClassName-[MethodName-]-PodName[-ContainerName].log".

So it will be easy to know which test class or method created which pod.

The directory where these logs are saved can be tweaked by using the
flag "logs.path". Default is "target/surefire-reports".