Add the ability to capture pods logs (#788)
* Remove unnecessary Session arguments on some methods

They are part of the class anyway.

* Kubernetes: Add the ability to capture pods logs

It might be useful for debugging purposes to have the ability
to automatically capture console log for all pods created
during the test execution.

It is disabled by default. To enable it, a new boolean flag is
added: "logs.copy".

The logs will be stored in the local filesystem, one file per pod.
Filenames will be "ClassName-[MethodName-]-PodName[-ContainerName].log".

So it will be easy to know which test class or method created which pod.

The directory where these logs are saved can be tweaked by using the
flag "logs.path". Default is "target/surefire-reports".

* Kubernetes: Add tests for the console log feature

* Kubernetes: Add the ability to capture cluster events

This is also useful for debugging, as complementary of getting
pods logs. If something goes wrong, users can check the kubernetes
events log.

This is enabled in conjunction with pods logs capture, using the same
set of flags.