Configure Arquillian Naming Service after process templates
This commit fixes the following issue:

If you define the following properties on arquillian.xml everything works as expected:

```
<property name="definitionsFile">src/test/resources/testrunner-pod.json</property>
<property name="proxiedContainerPorts">testrunner:9999,testrunner:9990,testrunner:8080</property>
```

The issue happens when there is no definitionsFile, the BuildablePodCube.java, the class where the routes are added in the naming service is not called.
And arquillian cube does not wait the environment get ready when asynchronous deployment is chosen.

With this fixes, all routes, even if there is multiple templates or there is not templates at all, just the definitions file, will be properly configured after all resources are in place.

In addition, fix the class TemplateProcessor, the following issue was happening when tryin to instantiate application-templates, line: Collections.copy(processedResources, resources);
java.lang.IndexOutOfBoundsException: Source does not fit in dest