Allow parallization of networks
If you're using networks and you want to parallelize your containers,
all the containers will end up on the same set of networks, which
could interfere with your test.

Even worse, as docker allows multiple networks with the same name, if
there's a problem with cleanup on a previous run, you may end up with
containers in different networks when you want them to be in the same
network.

This patch solves these problems by allowing the star pattern for
networks. It works the same as the container star pattern, and
networkMode and networks are updated within the container configuration.