Allow docker configuration to be loaded from the classpath
Allow users to specify the docker configuration file as a file on the
classpath, rather that on the filesystem. This means they won't have
to worry about absolute and relative paths, which can change in maven
depending on which flags are specified (forkCount for example).

It also allows configurations to be bundled into a jar and shared
between modules.